requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pydantic==2.10.4
pydantic-settings==2.4.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pycryptodome==3.20.0
email-validator==2.2.0
